Spine animation rendering, HaCreator for latest ms ver, & more fixes

+++ HaCreator/GUI/Repack.cs
@@ -27,20 +27,61 @@ namespace HaCreator.GUI
         public Repack()
         {
             InitializeComponent();
-            infoLabel.Text = "Files to repack:\r\n";
+
+            StringBuilder repackTxt = new StringBuilder("Files to repack:");
+            repackTxt.Append(Environment.NewLine);
+
             foreach (WzFile wzf in Program.WzManager.wzFiles.Values)
             {
                 if (Program.WzManager.wzFilesUpdated[wzf])
                 {
                     toRepack.Add(wzf);
-                    infoLabel.Text += wzf.Name + "\r\n";
+
+                    repackTxt.Append(wzf.Name);
+                    repackTxt.Append(Environment.NewLine);
                 }
+            }
+            infoLabel.Text = repackTxt.ToString();
+        }
+
+        /// <summary>
+        /// On closing form
+        /// </summary>
+        /// <param name="sender"></param>
+        /// <param name="e"></param>
+        private void Repack_FormClosing(object sender, FormClosingEventArgs e)
+        {
+            if (!button_repack.Enabled && !Program.Restarting)
+            { //Do not let the user close the form while saving
+                e.Cancel = true;
             }
         }
 
-        private void button1_Click(object sender, EventArgs e)
+        /// <summary>
+        /// Keydown
+        /// </summary>
+        /// <param name="sender"></param>
+        /// <param name="e"></param>
+        private void Repack_KeyDown(object sender, KeyEventArgs e)
         {
-            button1.Enabled = false;
+            if (e.KeyCode == Keys.Escape)
+            {
+                Close();
+            }
+            else if (e.KeyCode == Keys.Enter)
+            {
+                button_repack_Click(null, null);
+            }
+        }
+
+        /// <summary>
+        /// Repack button
+        /// </summary>
+        /// <param name="sender"></param>
+        /// <param name="e"></param>
+        private void button_repack_Click(object sender, EventArgs e)
+        {
+            button_repack.Enabled = false;
 
             Thread t = new Thread(new ThreadStart(RepackerThread));
             t.Start();
@@ -56,10 +97,22 @@ namespace HaCreator.GUI
             stateLabel.Text = state;
         }
 
-        private void FinishSuccess()
+        /// <summary>
+        /// On repacking completed
+        /// </summary>
+        /// <param name="bSaveFileInHaCreatorDirectory"></param>
+        private void FinishSuccess(bool bSaveFileInHaCreatorDirectory)
         {
-            MessageBox.Show("Repacked successfully, press OK to restart.");
-            Program.Restarting = true;
+            MessageBox.Show("Repacked successfully. " + (!bSaveFileInHaCreatorDirectory ? "press OK to restart." : "Please replace the files in HaCreator\\Output."));
+
+            if (!bSaveFileInHaCreatorDirectory)
+            {
+                Program.Restarting = true;
+            }
+            else
+            {
+                button_repack.Enabled = true;
+            }
             Close();
         }
 
@@ -68,7 +121,7 @@ namespace HaCreator.GUI
             Invoke((Action)delegate
             {
                 ChangeRepackState("ERROR While saving " + saveStage + ", aborted.");
-                button1.Enabled = true;
+                button_repack.Enabled = true;
                 ShowErrorMessage("There has been an error while saving, it is likely because you do not have permissions to the destination folder or the files are in use.\r\n\r\nPress OK to see the error details.");
                 ShowErrorMessage(e.Message + "\r\n" + e.StackTrace);
             });
@@ -76,20 +129,50 @@ namespace HaCreator.GUI
 
         private void RepackerThread()
         {
-            Invoke((Action)delegate { ChangeRepackState("Deleting old backups..."); });
+            Invoke((Action)delegate 
+            { 
+                ChangeRepackState("Deleting old backups..."); 
+            });
+
+            // Test for write access
+            string rootDir = Path.Combine(Program.WzManager.BaseDir, Program.APP_NAME);
+            string testDir = Path.Combine(rootDir, "Test");
+
+            bool bSaveFileInHaCreatorDirectory = false;
+            try
+            {
+                if (!Directory.Exists(testDir))
+                {
+                    Directory.CreateDirectory(testDir);
+                    Directory.Delete(testDir);
+                }
+            }
+            catch (Exception e)
+            {
+                if (e is UnauthorizedAccessException)
+                {
+                    bSaveFileInHaCreatorDirectory = true;
+                }
+            }
+            if (bSaveFileInHaCreatorDirectory)
+                rootDir = Path.Combine(Directory.GetCurrentDirectory(), Program.APP_NAME); 
+
             // Prepare directories
-            string rootDir = Path.Combine(Program.WzManager.BaseDir, "HaCreator");
             string backupDir = Path.Combine(rootDir, "Backup");
             string orgBackupDir = Path.Combine(rootDir, "Original");
             string XMLDir = Path.Combine(rootDir, "XML");
+
             try
             {
                 if (!Directory.Exists(backupDir))
                     Directory.CreateDirectory(backupDir);
+                
                 if (!Directory.Exists(orgBackupDir))
                     Directory.CreateDirectory(orgBackupDir);
+                
                 if (!Directory.Exists(XMLDir))
                     Directory.CreateDirectory(XMLDir);
+
                 foreach (FileInfo fi in new DirectoryInfo(backupDir).GetFiles())
                 {
                     fi.Delete();
@@ -103,7 +186,10 @@ namespace HaCreator.GUI
 
             // Save XMLs
             // We have to save XMLs first, otherwise the WzImages will already be disposed when we reach this code
-            Invoke((Action)delegate { ChangeRepackState("Saving XMLs..."); });
+            Invoke((Action)delegate 
+            { 
+                ChangeRepackState("Saving XMLs..."); 
+            });
             foreach (WzImage img in Program.WzManager.updatedImages)
             {
                 try
@@ -125,21 +211,51 @@ namespace HaCreator.GUI
             // Save WZ Files
             foreach (WzFile wzf in toRepack)
             {
-                Invoke((Action)delegate { ChangeRepackState("Saving " + wzf.Name + "..."); });
+                Invoke((Action)delegate 
+                { 
+                    ChangeRepackState("Saving " + wzf.Name + "..."); 
+                });
                 string orgFile = wzf.FilePath;
-                string tmpFile = orgFile + "$tmp";
+
+                string tmpFile;
+                if (!bSaveFileInHaCreatorDirectory)
+                    tmpFile = orgFile + "$tmp";
+                else
+                {
+                    string folderPath = Path.Combine(rootDir, "Output");
+                    tmpFile = Path.Combine(folderPath, wzf.Name);
+
+                    try
+                    {
+                        if (!Directory.Exists(folderPath))
+                            Directory.CreateDirectory(folderPath);
+
+                        if (!File.Exists(tmpFile))
+                            File.Create(tmpFile).Close();
+                    }
+                    catch (Exception e)
+                    {
+                        ShowErrorMessageThreadSafe(e, wzf.Name);
+                        return;
+                    }
+                }
+
                 try
                 {
                     wzf.SaveToDisk(tmpFile);
                     wzf.Dispose();
-                    string buPath = Path.Combine(orgBackupDir, Path.GetFileName(orgFile));
-                    // Try to backup to /Originals/ First, if there is already a file there, we are not original, so just backup to /Backup/
-                    if (File.Exists(buPath))
+
+                    if (!bSaveFileInHaCreatorDirectory) // only replace the original file if its saving in the maplestory folder
                     {
-                        buPath = Path.Combine(backupDir, Path.GetFileName(orgFile));
+                        string buPath = Path.Combine(orgBackupDir, Path.GetFileName(orgFile));
+                        // Try to backup to /Originals/ First, if there is already a file there, we are not original, so just backup to /Backup/
+                        if (File.Exists(buPath))
+                        {
+                            buPath = Path.Combine(backupDir, Path.GetFileName(orgFile));
+                        }
+                        File.Move(orgFile, buPath);
+                        File.Move(tmpFile, orgFile);
                     }
-                    File.Move(orgFile, buPath);
-                    File.Move(tmpFile, orgFile);
                 }
                 catch (Exception e)
                 {
@@ -148,27 +264,11 @@ namespace HaCreator.GUI
                 }
             }
 
-            Invoke((Action)delegate { ChangeRepackState("Finished"); FinishSuccess();});
-        }
-
-        private void Repack_FormClosing(object sender, FormClosingEventArgs e)
-        {
-            if (!button1.Enabled && !Program.Restarting)
-            { //Do not let the user close the form while saving
-                e.Cancel = true;
-            }
-        }
-
-        private void Repack_KeyDown(object sender, KeyEventArgs e)
-        {
-            if (e.KeyCode == Keys.Escape)
-            {
-                Close();
-            }
-            else if (e.KeyCode == Keys.Enter)
-            {
-                button1_Click(null, null);
-            }
+            Invoke((Action)delegate 
+            { 
+                ChangeRepackState("Finished"); 
+                FinishSuccess(bSaveFileInHaCreatorDirectory);
+            });
         }
     }
 }
